// front/src/app/api-client/model/meal-request.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Création / modification d\'un repas : une recette (recetteId) OU un libellé libre, pas les deux
 */
export interface MealRequest { 
    date: string;
    creneau: MealRequest.CreneauEnum;
    portions: number;
    /**
     * Recette type (exclusif avec libelle)
     */
    recetteId?: number;
    /**
     * Libellé libre, ex. « Restes », « Resto » (exclusif avec recetteId)
     */
    libelle?: string;
}
export namespace MealRequest {
    export const CreneauEnum = {
        PetitDejeuner: 'PETIT_DEJEUNER',
        Midi: 'MIDI',
        Souper: 'SOUPER',
    } as const;
    export type CreneauEnum = typeof CreneauEnum[keyof typeof CreneauEnum];
}