// front/src/app/api-client/model/recipe-summary-dto.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Recette type, sans ingrédients ni instructions (listes, recherche)
 */
export interface RecipeSummaryDTO { 
    id?: number;
    nom?: string;
    description?: string | null;
    portions?: number;
    /**
     * En minutes
     */
    tempsPreparation?: number | null;
}